+++ packages/cli/src/modules/source-control.ee/__tests__/source-control.service.test.ts
@@ -13,6 +13,7 @@ import type { EventService } from '@/events/event.service';
 import type { SourceControlExportService } from '../source-control-export.service.ee';
 import type { SourceControlGitService } from '../source-control-git.service.ee';
 import type { SourceControlImportService } from '../source-control-import.service.ee';
+import type { SourceControlContextFactory } from '../source-control-context.factory';
 import type { SourceControlScopedService } from '../source-control-scoped.service';
 import { sourceControlFoldersExistCheck } from '../source-control-helper.ee';
 import type { ExportResult } from '../types/export-result';
@@ -48,6 +49,7 @@ describe('SourceControlService', () => {
 	);
 	const sourceControlImportService = mock<SourceControlImportService>();
 	const sourceControlExportService = mock<SourceControlExportService>();
+	const sourceControlContextFactory = mock<SourceControlContextFactory>();
 	const sourceControlScopedService = mock<SourceControlScopedService>();
 	const gitService = mock<SourceControlGitService>();
 	const eventService = mock<EventService>();
@@ -57,6 +59,7 @@ describe('SourceControlService', () => {
 		preferencesService,
 		sourceControlExportService,
 		sourceControlImportService,
+		sourceControlContextFactory,
 		sourceControlScopedService,
 		eventService, // event service
 		mockStatusService as any, // status service
