+++ packages/cli/src/modules/source-control.ee/source-control.service.ee.ts
@@ -34,6 +34,7 @@ import {
 	getDeletedResources,
 	getNonDeletedResources,
 } from './source-control-resource-helper';
+import { SourceControlContextFactory } from './source-control-context.factory';
 import { SourceControlScopedService } from './source-control-scoped.service';
 import { SourceControlStatusService } from './source-control-status.service.ee';
 import type { ImportResult } from './types/import-result';
@@ -63,6 +64,7 @@ export class SourceControlService {
 		private sourceControlPreferencesService: SourceControlPreferencesService,
 		private sourceControlExportService: SourceControlExportService,
 		private sourceControlImportService: SourceControlImportService,
+		private sourceControlContextFactory: SourceControlContextFactory,
 		private sourceControlScopedService: SourceControlScopedService,
 		private readonly eventService: EventService,
 		private readonly sourceControlStatusService: SourceControlStatusService,
@@ -302,7 +304,7 @@ export class SourceControlService {
 			throw new BadRequestError('Cannot push onto read-only branch.');
 		}
 
-		const context = await this.sourceControlScopedService.createContext(user);
+		const context = await this.sourceControlContextFactory.createContext(user);
 
 		let filesToPush: SourceControlledFile[] = options.fileNames.map((file) => {
 			const normalizedPath = normalizeAndValidateSourceControlledFilePath(
@@ -614,7 +616,7 @@ export class SourceControlService {
 		commit?: string;
 	}): Promise<IWorkflowToImport> {
 		await this.sanityCheck();
-		const context = await this.sourceControlScopedService.createContext(user);
+		const context = await this.sourceControlContextFactory.createContext(user);
 		switch (type) {
 			case 'workflow': {
 				if (typeof id === 'undefined') {
