+++ packages/cli/test/integration/environments/source-control-import.service.test.ts
@@ -38,6 +38,7 @@ import { nanoid } from 'nanoid';
 import fsp from 'node:fs/promises';
 
 import { SourceControlImportService } from '@/modules/source-control.ee/source-control-import.service.ee';
+import { SourceControlContextFactory } from '@/modules/source-control.ee/source-control-context.factory';
 import { SourceControlScopedService } from '@/modules/source-control.ee/source-control-scoped.service';
 import type { ExportableCredential } from '@/modules/source-control.ee/types/exportable-credential';
 import type { IWorkflowToImport } from '@/interfaces';
@@ -63,6 +64,7 @@ describe('SourceControlImportService', () => {
 	let workflowTagMappingRepository: WorkflowTagMappingRepository;
 	let workflowHistoryRepository: WorkflowHistoryRepository;
 	let workflowHistoryService: WorkflowHistoryService;
+	let sourceControlContextFactory: SourceControlContextFactory;
 	let sourceControlScopedService: SourceControlScopedService;
 
 	const cipher = mockInstance(Cipher);
@@ -82,6 +84,7 @@ describe('SourceControlImportService', () => {
 		workflowTagMappingRepository = Container.get(WorkflowTagMappingRepository);
 		workflowHistoryRepository = Container.get(WorkflowHistoryRepository);
 		workflowHistoryService = Container.get(WorkflowHistoryService);
+		sourceControlContextFactory = Container.get(SourceControlContextFactory);
 		sourceControlScopedService = Container.get(SourceControlScopedService);
 		service = new SourceControlImportService(
 			mock(),
@@ -102,6 +105,7 @@ describe('SourceControlImportService', () => {
 			mock(),
 			folderRepository,
 			mock<InstanceSettings>({ n8nFolder: '/some-path' }),
+			sourceControlContextFactory,
 			sourceControlScopedService,
 			workflowHistoryService,
 			mock(),
@@ -260,7 +264,7 @@ describe('SourceControlImportService', () => {
 
 		it('should show all remote workflows for instance admins', async () => {
 			const result = await service.getRemoteVersionIdsFromFiles(
-				await sourceControlScopedService.createContext(globalAdmin),
+				await sourceControlContextFactory.createContext(globalAdmin),
 			);
 
 			expect(new Set(result.map((r) => r.id))).toEqual(
@@ -278,7 +282,7 @@ describe('SourceControlImportService', () => {
 
 		it('should show all remote workflows for instance owners', async () => {
 			const result = await service.getRemoteVersionIdsFromFiles(
-				await sourceControlScopedService.createContext(globalOwner),
+				await sourceControlContextFactory.createContext(globalOwner),
 			);
 
 			expect(new Set(result.map((r) => r.id))).toEqual(
@@ -296,7 +300,7 @@ describe('SourceControlImportService', () => {
 
 		it('should return no remote workflows for instance members', async () => {
 			const result = await service.getRemoteVersionIdsFromFiles(
-				await sourceControlScopedService.createContext(globalMember),
+				await sourceControlContextFactory.createContext(globalMember),
 			);
 
 			expect(result).toBeEmptyArray();
@@ -304,7 +308,7 @@ describe('SourceControlImportService', () => {
 
 		it('should return only remote workflows that belong to team project', async () => {
 			const result = await service.getRemoteVersionIdsFromFiles(
-				await sourceControlScopedService.createContext(teamAdmin),
+				await sourceControlContextFactory.createContext(teamAdmin),
 			);
 
 			expect(new Set(result.map((r) => r.id))).toEqual(
@@ -373,7 +377,7 @@ describe('SourceControlImportService', () => {
 		describe('if user is an instance owner', () => {
 			it('should get all available workflows on the instance', async () => {
 				const versions = await service.getLocalVersionIdsFromDb(
-					await sourceControlScopedService.createContext(instanceOwner),
+					await sourceControlContextFactory.createContext(instanceOwner),
 				);
 
 				expect(new Set(versions.map((v) => v.id))).toEqual(
@@ -391,7 +395,7 @@ describe('SourceControlImportService', () => {
 		describe('if user is a project admin of a team project', () => {
 			it('should only get all available workflows from the team project', async () => {
 				const versions = await service.getLocalVersionIdsFromDb(
-					await sourceControlScopedService.createContext(projectAdmin),
+					await sourceControlContextFactory.createContext(projectAdmin),
 				);
 
 				expect(new Set(versions.map((v) => v.id))).toEqual(
@@ -403,7 +407,7 @@ describe('SourceControlImportService', () => {
 		describe('if user is a project member of a team project', () => {
 			it('should not get any workflows', async () => {
 				const versions = await service.getLocalVersionIdsFromDb(
-					await sourceControlScopedService.createContext(projectMember),
+					await sourceControlContextFactory.createContext(projectMember),
 				);
 
 				expect(versions).toBeEmptyArray();
@@ -530,7 +534,7 @@ describe('SourceControlImportService', () => {
 
 		it('should show all remote credentials for instance admins', async () => {
 			const result = await service.getRemoteCredentialsFromFiles(
-				await sourceControlScopedService.createContext(globalAdmin),
+				await sourceControlContextFactory.createContext(globalAdmin),
 			);
 
 			expect(new Set(result.map((r) => r.id))).toEqual(
@@ -548,7 +552,7 @@ describe('SourceControlImportService', () => {
 
 		it('should show all remote credentials for instance owners', async () => {
 			const result = await service.getRemoteCredentialsFromFiles(
-				await sourceControlScopedService.createContext(globalOwner),
+				await sourceControlContextFactory.createContext(globalOwner),
 			);
 
 			expect(new Set(result.map((r) => r.id))).toEqual(
@@ -566,7 +570,7 @@ describe('SourceControlImportService', () => {
 
 		it('should return no remote credentials for instance members', async () => {
 			const result = await service.getRemoteCredentialsFromFiles(
-				await sourceControlScopedService.createContext(globalMember),
+				await sourceControlContextFactory.createContext(globalMember),
 			);
 
 			expect(result).toBeEmptyArray();
@@ -574,7 +578,7 @@ describe('SourceControlImportService', () => {
 
 		it('should return only remote credentials that belong to team project', async () => {
 			const result = await service.getRemoteCredentialsFromFiles(
-				await sourceControlScopedService.createContext(teamAdmin),
+				await sourceControlContextFactory.createContext(teamAdmin),
 			);
 
 			expect(new Set(result.map((r) => r.id))).toEqual(
@@ -663,7 +667,7 @@ describe('SourceControlImportService', () => {
 
 		it('should get all available credentials on the instance, for an instance owner', async () => {
 			const versions = await service.getLocalCredentialsFromDb(
-				await sourceControlScopedService.createContext(instanceOwner),
+				await sourceControlContextFactory.createContext(instanceOwner),
 			);
 
 			expect(new Set(versions.map((v) => v.id))).toEqual(
@@ -673,7 +677,7 @@ describe('SourceControlImportService', () => {
 
 		it('should only get all available credentials from the team project, for a project admin', async () => {
 			const versions = await service.getLocalCredentialsFromDb(
-				await sourceControlScopedService.createContext(projectAdmin),
+				await sourceControlContextFactory.createContext(projectAdmin),
 			);
 
 			expect(new Set(versions.map((v) => v.id))).toEqual(
@@ -683,7 +687,7 @@ describe('SourceControlImportService', () => {
 
 		it('should not get any workflows, for a project member', async () => {
 			const versions = await service.getLocalCredentialsFromDb(
-				await sourceControlScopedService.createContext(projectMember),
+				await sourceControlContextFactory.createContext(projectMember),
 			);
 
 			expect(versions).toBeEmptyArray();
@@ -713,7 +717,7 @@ describe('SourceControlImportService', () => {
 			);
 
 			const credentials = await service.getLocalCredentialsFromDb(
-				await sourceControlScopedService.createContext(instanceOwner),
+				await sourceControlContextFactory.createContext(instanceOwner),
 			);
 
 			const globalCred = credentials.find((c) => c.id === globalCredential.id);
@@ -738,7 +742,7 @@ describe('SourceControlImportService', () => {
 			);
 
 			const credentials = await service.getLocalCredentialsFromDb(
-				await sourceControlScopedService.createContext(instanceOwner),
+				await sourceControlContextFactory.createContext(instanceOwner),
 			);
 
 			const cred = credentials.find((c) => c.id === credential.id);
@@ -749,7 +753,7 @@ describe('SourceControlImportService', () => {
 
 		it('should include required properties in returned credentials', async () => {
 			const credentials = await service.getLocalCredentialsFromDb(
-				await sourceControlScopedService.createContext(instanceOwner),
+				await sourceControlContextFactory.createContext(instanceOwner),
 			);
 
 			expect(credentials.length).toBeGreaterThan(0);
@@ -823,7 +827,7 @@ describe('SourceControlImportService', () => {
 
 		it('should get all available folders on the instance, for an instance owner', async () => {
 			const folders = await service.getLocalFoldersAndMappingsFromDb(
-				await sourceControlScopedService.createContext(instanceOwner),
+				await sourceControlContextFactory.createContext(instanceOwner),
 			);
 
 			expect(new Set(folders.folders.map((v) => v.id))).toEqual(
@@ -833,7 +837,7 @@ describe('SourceControlImportService', () => {
 
 		it('should only get all available folders from the team project, for a project admin', async () => {
 			const versions = await service.getLocalFoldersAndMappingsFromDb(
-				await sourceControlScopedService.createContext(projectAdmin),
+				await sourceControlContextFactory.createContext(projectAdmin),
 			);
 
 			expect(new Set(versions.folders.map((v) => v.id))).toEqual(
@@ -843,7 +847,7 @@ describe('SourceControlImportService', () => {
 
 		it('should not get any folders, for a project member', async () => {
 			const versions = await service.getLocalFoldersAndMappingsFromDb(
-				await sourceControlScopedService.createContext(projectMember),
+				await sourceControlContextFactory.createContext(projectMember),
 			);
 
 			expect(versions.folders).toBeEmptyArray();
@@ -974,7 +978,7 @@ describe('SourceControlImportService', () => {
 
 		it('should show all remote tags and all remote mappings for instance admins', async () => {
 			const result = await service.getRemoteTagsAndMappingsFromFile(
-				await sourceControlScopedService.createContext(globalAdmin),
+				await sourceControlContextFactory.createContext(globalAdmin),
 			);
 
 			expect(new Set(result.tags.map((r) => r.id))).toEqual(
@@ -985,7 +989,7 @@ describe('SourceControlImportService', () => {
 
 		it('should show all remote tags and all remote mappings for instance owners', async () => {
 			const result = await service.getRemoteTagsAndMappingsFromFile(
-				await sourceControlScopedService.createContext(globalOwner),
+				await sourceControlContextFactory.createContext(globalOwner),
 			);
 
 			expect(new Set(result.tags.map((r) => r.id))).toEqual(
@@ -996,7 +1000,7 @@ describe('SourceControlImportService', () => {
 
 		it('should return all remote tags and no remote mappings for instance members', async () => {
 			const result = await service.getRemoteTagsAndMappingsFromFile(
-				await sourceControlScopedService.createContext(globalMember),
+				await sourceControlContextFactory.createContext(globalMember),
 			);
 
 			expect(new Set(result.tags.map((r) => r.id))).toEqual(
@@ -1007,7 +1011,7 @@ describe('SourceControlImportService', () => {
 
 		it('should return all remote tags and only remote mappings for in scope team for team admin', async () => {
 			const result = await service.getRemoteTagsAndMappingsFromFile(
-				await sourceControlScopedService.createContext(teamAdmin),
+				await sourceControlContextFactory.createContext(teamAdmin),
 			);
 
 			expect(new Set(result.tags.map((r) => r.id))).toEqual(
@@ -1124,7 +1128,7 @@ describe('SourceControlImportService', () => {
 
 		it('should get all available tags and mappings on the instance, for an instance owner', async () => {
 			const result = await service.getLocalTagsAndMappingsFromDb(
-				await sourceControlScopedService.createContext(instanceOwner),
+				await sourceControlContextFactory.createContext(instanceOwner),
 			);
 
 			expect(new Set(result.tags.map((v) => v.id))).toEqual(new Set([...tags.map((w) => w.id)]));
@@ -1145,7 +1149,7 @@ describe('SourceControlImportService', () => {
 
 		it('should only get all available tags and only mappings from the team project, for a project admin', async () => {
 			const result = await service.getLocalTagsAndMappingsFromDb(
-				await sourceControlScopedService.createContext(projectAdmin),
+				await sourceControlContextFactory.createContext(projectAdmin),
 			);
 
 			expect(new Set(result.tags.map((v) => v.id))).toEqual(new Set([...tags.map((w) => w.id)]));
@@ -1169,7 +1173,7 @@ describe('SourceControlImportService', () => {
 
 		it('should get all available tags but no mappings, for a project member', async () => {
 			const result = await service.getLocalTagsAndMappingsFromDb(
-				await sourceControlScopedService.createContext(projectMember),
+				await sourceControlContextFactory.createContext(projectMember),
 			);
 
 			expect(new Set(result.tags.map((v) => v.id))).toEqual(new Set([...tags.map((w) => w.id)]));
